refactor(tui): split MCP header helpers
